// supabase/functions/_shared/database.types.ts
export type Database = {
  public: {
    Tables: {
// ...
      articles: {
        Row: {
          category_id: string | null
          content: string
          created_at: string
          id: string
          published_at: string | null
          title: string
          updated_at: string | null
        }
        Insert: {
          category_id?: string | null
          content: string
          created_at?: string
          id?: string
          published_at?: string | null
          title: string
          updated_at?: string | null
        }
        Update: {
          category_id?: string | null
          content?: string
          created_at?: string
          id?: string
          published_at?: string | null
          title?: string
          updated_at?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "articles_category_id_categories_id_fk"
            columns: ["category_id"]
            isOneToOne: false
            referencedRelation: "categories"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
      sources: {
        Row: {
          created_at: string;
          updated_at: string | null;
          id: string;
          label: string;
          link: string;
        };
        Insert: {
          created_at?: string;
          updated_at?: string | null;
          id?: string;
          label: string;
          link: string;
        };
        Update: {
          created_at?: string;
          updated_at?: string | null;
          id?: string;
          label?: string;
          link?: string;
        };
        Relationships: [
          {
            foreignKeyName: "article_sources_source_id_sources_id_fk"; 
            columns: ["id"];
            isOneToOne: false;
            referencedRelation: "article_sources";
            referencedColumns: ["source_id"];
          }
        ];
      }
      article_sources: {
        Row: {
          article_id: string;
          source_id: string;
          created_at: string;
        };
        Insert: {
          article_id: string;
          source_id: string;
          created_at?: string;
        };
        Update: {
          article_id?: string;
          source_id?: string;
          created_at?: string;
        };
        Relationships: [
          {
            foreignKeyName: "article_sources_article_id_articles_id_fk";
            columns: ["article_id"];
            isOneToOne: false;
            referencedRelation: "articles";
            referencedColumns: ["id"];
          },
          {
            foreignKeyName: "article_sources_source_id_sources_id_fk"; 
            columns: ["source_id"];
            isOneToOne: false;
            referencedRelation: "sources";
            referencedColumns: ["id"];
          }
        ];
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}